apps/payment-service/pyproject.toml
[project]
name = "payment-service"
version = "0.1.0"
description = ""
authors = [
    {name = "konrad",email = "konrad.jed@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-json-logger (>=3.3.0,<4.0.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "opentelemetry-instrumentation-flask (>=0.53b1,<0.54)",
    "flask (>=3.1.0,<4.0.0)",
    "ecs-logging (>=2.2.0,<3.0.0)"
]

[tool.poetry]
name = "payment_service"
version = "0.1.0"
description = "A simple payment API with tracing & logging"
packages = [{include = "app", from = "."}]

#[tool.poetry.scripts]
#start = "flask.cli:main"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.1.0"
requests = "^2.32.3"
python-dotenv = "^1.1.0"
python-json-logger = "^3.3.0"
psutil = "^7.0.0"
logging = "^2.2.0"

[tool.poetry.group.dev.dependencies]
opentelemetry-api = "^1.32.1"
opentelemetry-sdk = "^1.32.1"
opentelemetry-exporter-otlp = "^1.32.1"
opentelemetry-instrumentation = "^0.53b1"
opentelemetry-instrumentation-flask = "^0.53b1"
opentelemetry-instrumentation-requests = "^0.53b1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
